refactor(core): give worker_pool own module
Moves the `AsyncWorkerPool` class to its own module
for better code organization and reusability.

This change promotes modularity and reduces the size of
the orchestrator module.
